requirements.txt
torch==1.13.1
torchvision==0.14.1
scikit-learn==1.3.2
argparse==1.4.0
matplotlib==3.5.3
pandas==1.5.3
numpy==1.22.3
tqdm==4.67.0
Pillow==9.5.0
